Create SB.h
The following code is in assembling instructions in the SB (Single Byte) format. The funn3() function determines the function code (funct3) based on the mnemonic of the instruction. The S() function takes an assembly instruction, extracts its components (e.g., registers and immediate values), and generates the corresponding machine code in hexadecimal format. It is coded for SB instructions like conditional branches (beq, bne, blt, bge,bltu,bgeu). The next part of the code also convert binary and hexadecimal representations of machine code.